Customize selected indicator, background color and shadow color.
These extra appearance selectors give you more customisation while
keeping with the spirit of the project.

* Change background color
* Change height of selected indicator
* Change color of selected indicator
* Change color of shadow color

Setting the background view via the appearance APIs wouldn't work when
using a DKTabPageViewController inside a navigation controller that
pushed another DKTabPageViewController. I looked through Apple's APIs
and on GitHub and noticed that no one ever exposes a UIView as an
appearance selector. I'm guessing it's unsupported.

If this change is unacceptable then I suppose we could add it back as a
public property that people could customise. Wanted to get an opinion on
this before I committed to doing it.